backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.0.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
qdrant-client==1.6.9
openai==1.3.5
pydantic==2.5.0
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1
Pillow==10.1.0
langchain-openai==0.1.3
requests
pandas
numpy
pydub==0.25.1
mangum==0.17.0

# Google Drive Integration
google-auth>=2.25.2
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.110.0
cryptography>=41.0.0 